+++ lib/token.dart
@@ -4,4 +4,5 @@ import 'builder_helper.dart' as helper;
 import 'parser_builder.dart';
 
 part 'src/token/tokenize.dart';
+part 'src/token/tokenize_similar_tags.dart';
 part 'src/token/tokenize_tags.dart';